requirements.txt
certifi==2019.6.16
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.5
django-cors-headers==3.1.0
django-filter==2.2.0
django-heroku==0.3.1
djangorestframework==3.10.3
gunicorn==19.9.0
heroku==0.1.4
idna==2.8
Markdown==3.1.1
mysqlclient==1.4.5
psycopg2==2.8.3
PyQRCode==1.2.1
python-dateutil==1.5
python-http-client==3.2.1
pytz==2019.2
requests==2.22.0
sendgrid==6.1.0
sqlparse==0.3.0
urllib3==1.25.3
whitenoise==4.1.3
